services/requirements.txt
# ================================================
# USDCOP Trading System - Unified Dependencies
# All services: trading-api, analytics-api, bi-api, multi-model-api, pipeline-api, realtime-ingestion-v2
# ================================================

# Core API Framework
fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-multipart==0.0.12

# Database - Sync & Async
psycopg2-binary==2.9.9
asyncpg==0.29.0
sqlalchemy==2.0.23

# Data Processing
pandas==2.2.3
numpy==1.26.4

# HTTP Clients
requests==2.31.0
httpx==0.27.0

# Redis - Sync & Async
redis[asyncio]==5.0.1

# MinIO/S3 Storage (for pipeline-api)
minio==7.2.8
boto3==1.35.36
botocore==1.35.36

# Technical Analysis (for pipeline-api, requires system TA-Lib)
ta-lib==0.4.32

# Resilience & Monitoring (for realtime-ingestion-v2)
circuitbreaker==2.0.0
prometheus-client==0.19.0

# WebSocket Support (for multi-model-api)
websockets==12.0

# Server-Sent Events (for trading-api-multi-model SSE streaming)
sse-starlette==2.0.0

# Utilities
python-dateutil==2.8.2
pytz==2024.1
python-dotenv==1.0.0

# Trading Calendar - Colombian Holidays
colombian-holidays>=1.0.0
holidays>=0.37

# ================================================
# MLOps - Production Inference & Monitoring
# ================================================

# ONNX Runtime for optimized inference
onnxruntime==1.16.3

# Data Drift Detection
evidently==0.4.16

# Model Export (optional - for development)
# onnx==1.15.0

# Statistical Tests (for drift fallback)
scipy==1.11.4

# Configuration Management
pyyaml==6.0.1
